# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sage Oak App Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /apps/{id}/upvote:
    post:
      operationId: toggleUpvote
      tags: [feedback]
      summary: Toggle the logged-in user's upvote for an app
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: New upvote state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpvoteState"
  /apps/{id}/issues:
    post:
      operationId: reportIssue
      tags: [feedback]
      summary: Report an issue on an app
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues:
    get:
      operationId: listIssues
      tags: [feedback]
      summary: List issues across all apps
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, resolved, all]
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
  /issues/{id}:
    patch:
      operationId: updateIssue
      tags: [feedback]
      summary: Resolve or reopen an issue (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
    delete:
      operationId: deleteIssue
      tags: [feedback]
      summary: Delete an issue and its related activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    User:
      type: object
      required: [id, email, displayName, role, tags, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        displayName: { type: string }
        role:
          type: string
          enum: [admin, staff]
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string }
...
    RaciValue:
      type: string
...
    RaciBoardPerson:
      type: object
      required: [name, value]
      properties:
        name: { type: string }
        value:
          $ref: "#/components/schemas/RaciValue"
...
    UpvoteState:
      type: object
      required: [applicationId, upvoted, upvoteCount]
      properties:
        applicationId: { type: integer }
        upvoted: { type: boolean }
        upvoteCount: { type: integer }
    IssueInput:
      type: object
      required: [comment]
      properties:
        comment: { type: string, minLength: 1 }
    Issue:
      type: object
      required: [id, applicationId, appName, userId, reporterName, comment, status, createdAt, raci]
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        appName: { type: string }
        userId: { type: integer }
        reporterName: { type: string }
        comment: { type: string }
        status:
          type: string
          enum: [open, resolved]
        createdAt: { type: string }
        resolvedAt: { type: string, nullable: true }
        raci:
          type: array
          items:
            $ref: "#/components/schemas/RaciBoardPerson"
    IssueUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, resolved]
...